Fix handling of bad pointers in arrays
When refactoring this code, I lost a behavior where a pointer to an
object that isn't found is dropped from the results (I had been
leaving an `undefined` in its place).

There's one tricky wrinkle, which is that if the value in the array is
`null`, it should be retained (see parse-community/parse-server#2189 for
more details on why this is the expected behavior).