Make `includePath` parallel-load safe
With parallel include fetching, there will by multiple invocations of
`includePath` in flight at once. Because we had been replacing
`this.response` with the result of this function, we suffered from a
common concurrency problem where in between reading a value and writing
it back, another task had mutated it and we would overwrite their
changes.

I started by promoting `includePath` from a standalone function to a
method of RestQuery. This lets it mutate `this.response` directly in a
very tightly scoped block. By keeping that block within a single
synchronous step, we avoid deferring to the event loop, meaning that we
should be safe from the type of concurrency issue described above.